Refactor the test structure
