+++ tests/integration/security/rbac/v1/group/group_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package groupv1
+package group
 
 import (
 	"testing"
