+++ tests/integration/security/rbac/v1/group/testdata/istio-clusterrbacconfig.yaml.tmpl
