+++ tests/integration/security/rbac/v1/group/testdata/istio-group-list-rbac-rules.yaml.tmpl
