+++ tests/integration/security/rbac/v2/basic/basic_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v2
+package basic
 
 import (
 	"testing"
