+++ tests/integration/security/rbac/v2/basic/testdata/istio-clusterrbacconfig.yaml.tmpl
