+++ tests/integration/security/rbac/v2/basic/testdata/istio-rbac-v2-rules.yaml.tmpl
